daemon/pyproject.toml
[project]
name = "wallgardend"
version = "0.1.0"
description = "Wall garden control daemon — owns the hardware and the safety loop."
requires-python = ">=3.11"
readme = "../README.md"

dependencies = [
    "psycopg[binary,pool]>=3.2",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "structlog>=24.1",
    "pillow>=10.3",                # synthetic plant images for the camera mock
    "pyyaml>=6.0",
    "prometheus-client>=0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "hypothesis>=6.100",
    "freezegun>=1.5",
    "ruff>=0.5",
]
# Pi-only deps are NOT in pyproject.toml — they pull arch-specific wheels that
# fail to resolve on macOS. Install them on the Pi with:
#   uv pip install -r requirements-pi.txt
# The Pi backend lazy-imports them so the daemon loads fine without them on dev.

[project.scripts]
wallgardend = "wallgardend.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["wallgardend"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "S"]
ignore = ["S101"]  # allow asserts in tests
